backend/requirements.txt
flask
bcrypt
flask-sqlalchemy
flask-login
nose
mysql
